+++ app/assets/javascripts/student_profile/ElaDetails.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import {merge} from '../helpers/merge';
 import ProfileChart from './ProfileChart';
 import ProfileChartSettings from './ProfileChartSettings';
+import DetailsSection from './DetailsSection';
 
 /*
 This renders details about ELA performance and growth in the student profile page.
@@ -39,8 +40,8 @@ export default class ElaDetails extends React.Component {
       <div className="ElaDetails">
         {this.renderNavBar()}
         {this.renderStarReading()}
-        {this.renderMCASELAScores()}
         {this.renderMCASELANextGenScores()}
+        {this.renderMCASELAScores()}
         {this.renderMCASELAGrowth()}
       </div>
     );
@@ -95,21 +96,6 @@ export default class ElaDetails extends React.Component {
     );
   }
 
-  renderHeader(title) {
-    return (
-      <div style={styles.secHead}>
-        <h4 style={styles.title}>
-          {title}
-        </h4>
-        <span style={styles.navTop}>
-          <a href="#">
-            Back to top
-          </a>
-        </span>
-      </div>
-    );
-  }
-
   renderStarReading() {
     const {chartData} = this.props;
 
@@ -117,8 +103,7 @@ export default class ElaDetails extends React.Component {
     const starReading = chartData.star_series_reading_percentile.reverse();
 
     return (
-      <div id='Star' style={styles.container}>
-        {this.renderHeader('STAR Reading, last 4 years')}
+      <DetailsSection anchorId="Star" title="STAR Reading, last 4 years">
         <ProfileChart
           quadSeries={[{ name: 'Percentile rank', data: starReading }]}
           titleText=''
@@ -127,7 +112,7 @@ export default class ElaDetails extends React.Component {
             title: { text: 'Percentile rank' }
           })}
           showGradeLevelEquivalent= { true }/>
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -137,8 +122,7 @@ export default class ElaDetails extends React.Component {
     if (!data || data.length === 0) return null;
 
     return (
-      <div id="NextGenScores" style={styles.container}>
-        {this.renderHeader('MCAS Next Gen ELA Scores, last 4 years')}
+      <DetailsSection anchorId="NextGenScores" title="MCAS Next Gen ELA Scores, last 4 years">
         <ProfileChart
           quadSeries={[{
             name: 'Scaled score',
@@ -153,7 +137,7 @@ export default class ElaDetails extends React.Component {
             title: { text: 'Scaled score' }
           })} />
 
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -163,8 +147,7 @@ export default class ElaDetails extends React.Component {
     if (!data || data.length === 0) return null;
 
     return (
-      <div id="Scores" style={styles.container}>
-        {this.renderHeader('MCAS ELA Scores, last 4 years')}
+      <DetailsSection anchorId="Scores" title="MCAS ELA Scores, last 4 years">
         <ProfileChart
           quadSeries={[{
             name: 'Scaled score',
@@ -176,14 +159,13 @@ export default class ElaDetails extends React.Component {
             plotLines: ProfileChartSettings.mcas_level_bands,
             title: { text: 'Scaled score' }
           })} />
-      </div>
+      </DetailsSection>
     );
   }
 
   renderMCASELAGrowth() {
     return (
-      <div id="SGPs" style={styles.container}>
-        {this.renderHeader('Student growth percentile (SGP), last 4 years')}
+      <DetailsSection anchorId="SGPs" title="Student growth percentile (SGP), last 4 years">
         <ProfileChart
           quadSeries={[{
             name: '',
@@ -194,7 +176,7 @@ export default class ElaDetails extends React.Component {
           yAxis={merge(this.percentileYAxis(), {
             title: { text: 'Student growth percentile (SGP)' }
           })} />
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -212,31 +194,7 @@ ElaDetails.propTypes = {
 };
 
 const styles = {
-  title: {
-    color: 'black',
-    paddingBottom: 20,
-    fontSize: 24
-  },
-  container: {
-    width: '100%',
-    marginTop: 50,
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    border: '1px solid #ccc',
-    padding: '30px 30px 30px 30px',
-    position: 'relative'
-  },
-  secHead: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    position: 'relative',
-    bottom: 10
-  },
   navBar: {
     fontSize: 18
-  },
-  navTop: {
-    textAlign: 'right',
-    verticalAlign: 'text-top'
   }
 };
