+++ app/assets/javascripts/student_profile/MathDetails.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import {merge} from '../helpers/merge';
 import ProfileChart from './ProfileChart';
 import ProfileChartSettings from './ProfileChartSettings';
+import DetailsSection from './DetailsSection';
 
 
 /*
@@ -41,8 +42,8 @@ export default class MathDetails extends React.Component {
       <div className="MathDetails">
         {this.renderNavBar()}
         {this.renderStarMath()}
-        {this.renderMCASMathScores()}
         {this.renderMCASMathNextGenScores()}
+        {this.renderMCASMathScores()}
         {this.renderMCASMathGrowth()}
       </div>
     );
@@ -97,21 +98,6 @@ export default class MathDetails extends React.Component {
     );
   }
 
-  renderHeader(title) {
-    return (
-      <div style={styles.secHead}>
-        <h4 style={styles.title}>
-          {title}
-        </h4>
-        <span style={styles.navTop}>
-          <a href="#">
-            Back to top
-          </a>
-        </span>
-      </div>
-    );
-  }
-
   renderStarMath() {
     const {chartData} = this.props;
 
@@ -119,8 +105,7 @@ export default class MathDetails extends React.Component {
     const starMathPercentile = chartData.star_series_math_percentile.reverse();
 
     return (
-      <div id="starMath" style={styles.container}>
-        {this.renderHeader('STAR Math, last 4 years')}
+      <DetailsSection anchorId="starMath" title="STAR Math, last 4 years">
         <ProfileChart
           quadSeries={[{ name: 'Percentile rank', data: starMathPercentile }]}
           titleText="STAR Math, last 4 years"
@@ -129,7 +114,7 @@ export default class MathDetails extends React.Component {
             title: { text: 'Percentile rank' }
           })}
           showGradeLevelEquivalent= { true }/>
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -139,8 +124,7 @@ export default class MathDetails extends React.Component {
     if (!data || data.length === 0) return null;
 
     return (
-      <div id="MCASMathNextGen" style={styles.container}>
-        {this.renderHeader('MCAS Next Gen Math Scores')}
+      <DetailsSection anchorId="MCASMathNextGen" title="MCAS Next Gen Math Scores">
         <ProfileChart
           quadSeries={[{
             name: 'Scaled score',
@@ -154,7 +138,7 @@ export default class MathDetails extends React.Component {
             plotLines: ProfileChartSettings.mcas_next_gen_level_bands,
             title: { text: 'Scaled score' }
           })} />
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -164,8 +148,7 @@ export default class MathDetails extends React.Component {
     if (!data || data.length === 0) return null;
 
     return (
-      <div id="MCASMath" style={styles.container}>
-        {this.renderHeader('MCAS Math Scores, last 4 years')}
+      <DetailsSection anchorId="MCASMath" title="MCAS Math Scores, last 4 years">
         <ProfileChart
           quadSeries={[{
             name: 'Scaled score',
@@ -177,14 +160,13 @@ export default class MathDetails extends React.Component {
             plotLines: ProfileChartSettings.mcas_level_bands,
             title: { text: 'Scaled score' }
           })} />
-      </div>
+      </DetailsSection>
     );
   }
 
   renderMCASMathGrowth() {
     return (
-      <div id="MCASMathGrowth" style={styles.container}>
-        {this.renderHeader('MCAS Math SGPs, last 4 years')}
+      <DetailsSection anchorId="MCASMathGrowth" title="MCAS Math SGPs, last 4 years">
         <ProfileChart
           quadSeries={[{
             name: 'Student growth percentile (SGP)',
@@ -195,7 +177,7 @@ export default class MathDetails extends React.Component {
           yAxis={merge(this.percentileYAxis(), {
             title: { text: 'Student growth percentile (SGP)' }
           })} />
-      </div>
+      </DetailsSection>
     );
   }
 
@@ -213,31 +195,7 @@ MathDetails.propTypes = {
 };
 
 const styles = {
-  title: {
-    color: 'black',
-    paddingBottom: 20,
-    fontSize: 24
-  },
-  container: {
-    width: '100%',
-    marginTop: 50,
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    border: '1px solid #ccc',
-    padding: '30px 30px 30px 30px',
-    position: 'relative'
-  },
-  secHead: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    position: 'relative',
-    bottom: 10
-  },
   navBar: {
     fontSize: 18
-  },
-  navTop: {
-    textAlign: 'right',
-    verticalAlign: 'text-top'
   }
 };
